requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
PyMuPDF==1.25.4
python-dotenv==1.1.0
scikit-learn==1.6.1
spacy==3.8.4
transformers==4.51.1
torch==2.6.0
sentence-transformers==4.0.1
whitenoise==6.9.0
langchain==0.3.22
langchain-community==0.3.20
faiss-cpu==1.10.0

# Install spaCy model
python -m spacy download en_core_web_sm
